Minor modifs

+++ admin/install.php
@@ -79,7 +79,7 @@ function bad_tournament_install_init(){
                   lastname varchar(150) NOT NULL,
                   player_level int(6) unsigned NOT NULL,
                   status int(6) unsigned NOT NULL,
-                  club_id int(6) unsigned NOT NULL,
+                  club_id bigint(20) unsigned NOT NULL,
                   player_id varchar(50) NOT NULL,
                   birtdate date,
                   sex int(2) unsigned NOT NULL,
